// types/periodButtons.ts
export enum PeriodDisplay {
  oneDay = '1D',
  oneWeek = '1W',
  oneMonth = '1M',
  // threeMonth = '3m',
  // sixMonth = '6m',
  oneYear = '1Y',
}
export enum Period {
  oneDay = 'DAY',
  oneWeek = 'WEEK',
  oneMonth = 'MONTH',
  // threeMonth = '3MONTHS',
  // sixMonth = '6MONTHS',
  oneYear = 'YEAR',
}